+++ backend/sokdak/src/main/java/com/wooteco/sokdak/member/dto/EmailRequest.java
@@ -1,10 +1,12 @@
 package com.wooteco.sokdak.member.dto;
 
+import javax.validation.constraints.NotBlank;
 import lombok.Getter;
 
 @Getter
 public class EmailRequest {
 
+    @NotBlank
     private String email;
 
     public EmailRequest() {
