+++ backend/sokdak/src/test/java/com/wooteco/sokdak/auth/controller/AuthControllerTest.java
@@ -14,6 +14,8 @@ import com.wooteco.sokdak.auth.dto.AuthInfo;
 import com.wooteco.sokdak.auth.dto.LoginRequest;
 import com.wooteco.sokdak.auth.exception.LoginFailedException;
 import com.wooteco.sokdak.member.domain.RoleType;
+import com.wooteco.sokdak.member.dto.VerificationRequest;
+import com.wooteco.sokdak.member.exception.InvalidAuthCodeException;
 import com.wooteco.sokdak.util.ControllerTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -84,5 +86,24 @@ class AuthControllerTest extends ControllerTest {
                 .assertThat()
                 .apply(document("refresh/success"))
                 .statusCode(HttpStatus.NO_CONTENT.value());
+    }
+
+    @DisplayName("인증코드가 일치하지 않거나 만료되었으면 400 반환")
+    @Test
+    void verifyAuthCode_Exception_different() {
+        VerificationRequest verificationRequest = new VerificationRequest("test@gmail.com", "a1b2c3");
+        doThrow(new InvalidAuthCodeException())
+                .when(authService)
+                .verifyAuthCode(any());
+
+        restDocs
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(verificationRequest)
+                .when().post("/members/signup/email/verification")
+                .then().log().all()
+                .assertThat()
+                .body("message", equalTo("잘못된 인증번호입니다."))
+                .apply(document("member/verification/fail"))
+                .statusCode(HttpStatus.BAD_REQUEST.value());
     }
 }
